Refs #29544 move wire positions to utils module
